questions/daily_temperatures.py
"""
Problem: Days to Wait for Warmer Temperature

Description:
Given a list of daily temperatures, the task is to determine how many days one would have to wait until experiencing a warmer temperature. If there's no warmer temperature in the future, mark the day with 0.

Function Signature:
def days_to_wait_for_warmer(temperatures: List[int]) -> List[int]:

Inputs:
    - temperatures (List[int]): A list of daily temperatures.

Returns:
    - List[int]: A list indicating the number of days to wait for a warmer temperature for each corresponding day in the input.

Notes:
    - The returned list should have the same length as the input list.
    - A warmer temperature is any temperature higher than the current day's temperature.

Examples:

1. Input: [70, 73, 75, 71, 69, 72, 76, 73]
   Output: [1, 1, 4, 2, 1, 1, 0, 0]
   Explanation: For the first temperature, 70, the next warmer temperature is 73, which is 1 day away.

2. Input: [73, 72, 71, 70]
   Output: [0, 0, 0, 0]
   Explanation: There are no warmer temperatures for any of the days.

3. Input: [70, 71, 72, 73]
   Output: [1, 1, 1, 0]
   Explanation: Each day has a warmer temperature the following day, except for the last day.

Hints:
    - Consider using a stack to keep track of temperatures and their indices.
    - Iterate through the temperatures, and for each temperature, check if it's warmer than the last one in the stack.

Tags:
    - Stack
"""

from typing import List

def days_to_wait_for_warmer(temperatures: List[int]) -> List[int]:
    # TODO: Implement the function
    return [0]
